Extract test fixture helper and configure package files
Move the fixture utility into its own module for reusability
and add a `files` field to package.json to limit published artifacts.
